Error suppression (#4)
This adds a model of error suppression. Rather than report a warning any
time there's a failure of subtyping, we suppress the error in the case
that a supertype of `error` was involved. In particular, `error <: any`,
so we suppress errors that involve `any`. For example, if `f` has type
`(number -> any) -> number` and `g` has type `number -> string` then we
suppress the error in the function application `f(g)`.

To get type safety back, we introduce a notion of "unsafe" type (ones
mentioning `error` or `any`) and show that for safe programs (that is,
ones which only use safe types), we get the usual type safety result.
This works because any time an error is suppressed, the program is
unsafe.